pyproject.toml
[tool.poetry]
name = "fullstack todo"
version = "0.1.0"
description = "FullStack ToDo application"
authors = ["Aron Sadegh  <aronesadegh@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.110.2"
uvicorn = "^0.29.0"
sqlalchemy = "^2.0.29"
pydantic = {extras = ["email"], version = "^2.7.1"}
python-multipart = "^0.0.9"
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
pymysql = "^1.1.0"
passlib = {extras = ["cryptography"], version = "^1.7.4"}
alembic = "^1.13.1"
jinja2 = "^3.1.3"


[tool.poetry.group.dev.dependencies]
pytest = "^8.2.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
